Moved value getting and setting into functions
Cleaned up a lot of the duplicated functionality inside each function, so now we're moving the value getting into a single function to add support for mode parameter modes, and reduce duplicated lines.

Now the Instruction handler figures out which function to run, and each function handles grabbing the parameters that they need instead of splitting that around.